refactor: adapter 제거
